Create top-level catalogs module
