+++ api/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod catalogs;
 pub mod conf;
 pub mod db;
 pub mod filters;
