+++ api/src/routes/catalogs.rs
@@ -1,4 +1,6 @@
 use crate::{
+    catalogs::catalog_exists,
+    db::PROTECTED_COLLECTION_NAMES,
     filters::{parse_filter, parse_optional_filter},
     models::response,
 };
@@ -12,29 +14,6 @@ use std::fmt;
 use utoipa::openapi::RefOr;
 use utoipa::openapi::schema::{ObjectBuilder, Schema};
 use utoipa::{PartialSchema, ToSchema};
-
-/// Protected names for operational data collections, which should not be used
-/// for analytical data catalogs
-const PROTECTED_COLLECTION_NAMES: [&str; 2] = ["users", "filters"];
-
-/// Check if a catalog exists
-pub async fn catalog_exists(db: &Database, catalog_name: &str) -> bool {
-    if catalog_name.is_empty() {
-        return false; // Empty catalog names are not valid
-    }
-    if PROTECTED_COLLECTION_NAMES.contains(&catalog_name) {
-        return false; // Protected names cannot be used as catalog names
-    }
-    // Get collection names in alphabetical order
-    let collection_names = match db.list_collection_names().await {
-        Ok(c) => c,
-        Err(_) => return false,
-    };
-    // Convert catalog name to collection name
-    let collection_name = catalog_name.to_string();
-    // Check if the collection exists
-    collection_names.contains(&collection_name)
-}
 
 #[derive(serde::Deserialize)]
 struct CatalogsQueryParams {
